TaskResultDao add support for size and contains methods

+++ taskModel/src/main/java/storage/InMemoryTasksResultStorage.java
@@ -14,7 +14,7 @@ public class InMemoryTasksResultStorage implements TasksResultsStorage {
   }
 
   @Override
-  public int size() {
+  public long size() {
     return taskResults.size();
   }
